+++ cache/geopoints/utils/GeoHashWrapper.cpp
@@ -2,9 +2,9 @@
 #include <cstdint>
 
 extern "C" {
-    uint64_t geohash_int(double latitude, double longitude, unsigned int precision);
+    uint64_t geohash_int(float latitude, float longitude, unsigned int precision);
 }
 
-uint64_t calculate_geohash(double latitude, double longitude) {
+uint64_t calculate_geohash(float latitude, float longitude) {
     return geohash_int(latitude, longitude, 12);
 }
